requirements.txt
flask==3.0.0
flask-restful==0.3.10
flask-pymongo==2.3.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.5.3
pytest==7.4.0
requests==2.31.0
gunicorn==21.2.0
